Plumb a clip bounding box around through overzoom
